Updating the computations in the data, and updating accordingly the other scripts
Only some bunches are dissected, so it is not right to e.g. use the sum of the fruits biomass as the total biomass of fruits harvested for a month, as we would miss all bunches that were not dissected. We now compute everything at the bunch scale, and updated the computations of the number of bunches, and also added the dry mass of the whole bunch based on an average value of the water content from the dissected bunch for the trees of the same progeny and MAP.